+++ src/server/message_handler/diagnostic/mod.rs
@@ -1,3 +1,5 @@
+mod auto_fix_prefixes;
+pub mod duplicate_prefix_declaration;
 pub mod invalid_projection_variable;
 pub mod same_subject;
 pub mod syntax_error;
@@ -6,22 +8,14 @@ pub mod undeclared_prefix;
 pub mod ungrouped_select_variable;
 pub mod unused_prefix_declaration;
 
-use super::code_action::remove_prefix_declaration;
 use crate::server::{
     Server,
-    lsp::{
-        DiagnosticRequest, DiagnosticResponse, WorkspaceEditRequest, base_types::LSPAny,
-        diagnostic::Diagnostic, errors::LSPError,
-    },
-    message_handler::code_action::declare_prefix,
+    lsp::{DiagnosticRequest, DiagnosticResponse, errors::LSPError},
 };
+use auto_fix_prefixes::{auto_fix_prefixes, client_support_workspace_edits};
 use futures::lock::Mutex;
 use ll_sparql_parser::ast::{AstNode, QueryUnit};
-use std::{
-    collections::{HashMap, HashSet},
-    convert::identity,
-    rc::Rc,
-};
+use std::rc::Rc;
 
 #[tracing::instrument(skip_all, fields(id = %request.get_id(), uri = %request.params.text_document.uri))]
 pub(super) async fn handle_diagnostic_request(
@@ -41,99 +35,25 @@ pub(super) async fn handle_diagnostic_request(
         "diagnostics are currently only supported for query operations",
     ))?;
     let mut diagnostic_accu = Vec::new();
-    macro_rules! add {
+    macro_rules! add_diagnostic {
         ($diagnostic_provider:path) => {
             if let Some(diagnostics) = $diagnostic_provider(document, &ast, &server) {
                 diagnostic_accu.extend(diagnostics);
             }
         };
     }
-    add!(unused_prefix_declaration::diagnostics);
-    add!(undeclared_prefix::diagnostics);
-    add!(uncompacted_uri::diagnostics);
-    add!(ungrouped_select_variable::diagnostics);
-    add!(invalid_projection_variable::diagnostics);
-    add!(same_subject::diagnostics);
-    add!(syntax_error::diagnostics);
+    add_diagnostic!(unused_prefix_declaration::diagnostics);
+    add_diagnostic!(undeclared_prefix::diagnostics);
+    add_diagnostic!(uncompacted_uri::diagnostics);
+    add_diagnostic!(ungrouped_select_variable::diagnostics);
+    add_diagnostic!(invalid_projection_variable::diagnostics);
+    add_diagnostic!(same_subject::diagnostics);
+    add_diagnostic!(syntax_error::diagnostics);
+    add_diagnostic!(duplicate_prefix_declaration::diagnostics);
 
     if client_support_workspace_edits(&server) {
-        declare_and_undeclare_prefixes(&mut server, &request, &diagnostic_accu);
+        auto_fix_prefixes(&mut server, &request, &diagnostic_accu);
     }
 
     server.send_message(DiagnosticResponse::new(request.get_id(), diagnostic_accu))
-}
-
-fn declare_and_undeclare_prefixes(
-    server: &mut Server,
-    request: &DiagnosticRequest,
-    diagnostics: &[Diagnostic],
-) {
-    let document_uri = request.params.text_document.uri.clone();
-    let mut prefixes = HashSet::<&str>::new();
-    let edits: Vec<_> = diagnostics
-        .iter()
-        .filter_map(|diagnostic| {
-            if let Some(LSPAny::String(prefix)) = diagnostic.data.as_ref() {
-                if prefixes.insert(prefix) {
-                    match diagnostic.code.as_ref() {
-                        Some(code)
-                            if code == &*undeclared_prefix::CODE
-                                && server.settings.prefixes.as_ref().is_some_and(|prefixes| {
-                                    prefixes.add_missing.is_some_and(identity)
-                                }) =>
-                        {
-                            declare_prefix(server, &document_uri, diagnostic.clone())
-                        }
-                        Some(code)
-                            if code == &*unused_prefix_declaration::CODE
-                                && server.settings.prefixes.as_ref().is_some_and(|prefixes| {
-                                    prefixes.remove_unused.is_some_and(identity)
-                                }) =>
-                        {
-                            remove_prefix_declaration(server, &document_uri, diagnostic.clone())
-                        }
-                        _ => Ok(None),
-                    }
-                    .ok()
-                    .flatten()
-                    .and_then(|code_action| code_action.edit.changes)
-                    .and_then(|mut changes| changes.remove(&document_uri))
-                } else {
-                    None
-                }
-            } else {
-                None
-            }
-        })
-        .flatten()
-        .collect();
-    if !edits.is_empty() {
-        let request_id = server.bump_request_id();
-        if let Err(err) = server.send_message(WorkspaceEditRequest::new(
-            request_id,
-            HashMap::from_iter([(document_uri, edits)]),
-        )) {
-            tracing::error!("Sending \"workspace/applyEdit\" request failed:\n{:?}", err);
-        }
-    }
-}
-
-fn client_support_workspace_edits(server: &Server) -> bool {
-    server
-        .client_capabilities
-        .as_ref()
-        .is_some_and(|client_capabilities| {
-            client_capabilities
-                .workspace
-                .as_ref()
-                .and_then(|workspace_capabilities| workspace_capabilities.apply_edit)
-                .is_some_and(|flag| flag)
-                && client_capabilities
-                    .workspace
-                    .as_ref()
-                    .and_then(|workspace_capabilities| {
-                        workspace_capabilities.workspace_edit.as_ref()
-                    })
-                    .is_some_and(|capability| capability.document_changes.is_some_and(|flag| flag))
-        })
 }